Changes package location of CLIJxFilterOp Puts Util class as an inner class of CLIJxFilterOp Removes powerpool method in CLIJxPool Adds self contained DemoDummyFiltering Moves Lazy class as inner class of DemoDummyFiltering Removes N5 dependencies and associated demo and classes pom.xml: - puts SNAPSHOT suffix in version - comment scijava app directory since it is dev specific - removes non working N5 dependencies - removes unused picocli dependecy - put bdv-vistools dependency scope as test Ignores tests since they are highly dependent on the hardware of the developer Removes uncompiling demo depending on N5
